mmview: Fix concurrency bug with mm_users
We allowed mm_users to temporarily drop to zero without freeing the mm in
case a view is not currently used but is still available (determined by the
MMVIEW_AVAILABLE flag). Migration to such a view is possible which means
mm_users can be incremented again. Therefore, we made mmput additionally
check MMVIEW_AVAILABLE before freeing the mm. This results in a bug when
the flag and mm_users get modified during a concurrent mmput.

Fix this by counting the MMVIEW_AVAILABLE flag as an additional mm user.
This removes the need of explicitly checking the flag in mmview_put. The
common.users counter still only counts the "real" users of all views.